feat: Move PromptPartSelection to `types.py` for 3.3
